requirements.txt
flask==3.0.3
waitress==3.0.0
python-dotenv==1.0.1
torch==2.4.0
torchvision==0.19.0
numpy<2
pillow==10.4.0
opencv-python==4.10.0.84
matplotlib==3.9.2
scikit-learn==1.5.1
pytest==8.3.2
requests==2.32.3
segmentation-models-pytorch==0.3.4
albumentations==1.4.14
tqdm==4.66.5
pandas==2.2.2